+++ examples/run_imgpath_correct.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 cd ..
-python nmag/md_manager/imgpath_corrector.py --input_folder examples/imgpath_correct --output_folder examples/imgpath_correct_new
+python nmag/mmage/imgpath_corrector.py --input_folder examples/imgpath_correct --output_folder examples/imgpath_correct_new