+++ examples/run_pdf2md.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 cd ..
-python nmag/pdf_manager/pdf2md.py --input_folder examples/pdf2md --output_folder examples/pdf2md
+python nmag/pmage/pdf2md.py --input_folder examples/pdf2md --output_folder examples/pdf2md