+++ CHANGELOG.md
@@ -2,6 +2,6 @@
 
 ### Feat
 
-- add setup.py and commitizen for changelog automation and move img_concat to img_manager folder, add a skeleton for cli
+- add setup.py and commitizen for changelog automation and move img_concat to image folder, add a skeleton for cli
 - add output_folder input arg in renamer
 - add args parser in both typora functions